Change pattern for executables in .gitignore & Add WardShip XML files
The '.gitignore' contained a pattern that matched any "bin/" directories in the
repository. This pattern has been removed and the pattern "*.class" has been added. This
is to retain a bit more of the directory structure for my projects when I push changes
to github and then pull them in the Windows side of my system.

This commit also adds the eclipse '.classpath' and '.project' configuration files for
the 'WardShip' project in 'Module7'.